+++ src/logging/formatting.js
